requirements.txt
openai>=1.0
google-generativeai>=0.8
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scipy>=1.10
